adds more styled components to individual folders
