+++ src/components/Common/QuestionForm/index.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+const QuestionForm = styled.form`
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: center;
+  justify-content: center;
+  @media (max-width: 400px) {
+    flex-direction: column;
+  }
+`;
+
+export default QuestionForm;
