+++ src/components/Common/StripeButton/index.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+import PrimaryButton from "../PrimaryButton";
+
+const StripeButton = styled(PrimaryButton)`
+  background: #fabc09;
+  color: #ffffff;
+  a {
+    text-decoration: none;
+    color: white;
+  }
+`;
+
+export default StripeButton;
