+++ src/components/Common/StyledScrollPanel/index.js
@@ -1,0 +1,13 @@
+import { ScrollPanel } from "primereact/scrollpanel";
+import styled from "styled-components";
+
+const StyledScrollPanel = styled(ScrollPanel)`
+  &&& {
+    width: 100%;
+    margin-bottom: 20px;
+    border: none;
+    height: 75px;
+  }
+`;
+
+export default StyledScrollPanel;
