+++ src/components/Dashboard/index.js
@@ -40,7 +40,7 @@ const ButtonNavLink = styled(Link)`
   cursor: pointer;
 `;
 
-class MeetingList extends Component {
+class Dashboard extends Component {
   componentDidMount() {
     if (this.props.token && !localStorage.getItem("jwt")) {
       this.props.callGoogleLogIn(history, this.props.token);
@@ -96,4 +96,4 @@ const mapStateToProps = state => {
 export default connect(
   mapStateToProps,
   { getMeetings, callDeleteMeeting, callGoogleLogIn }
-)(MeetingList);
+)(Dashboard);
