+++ src/components/UpdateMeeting/index.js
@@ -338,8 +338,7 @@ class UpdateMeeting extends Component {
                   onChange={this.changeHandler}
                   inputClassName="input"
                   className="datePicker"
-                  placeholder="Start"
-                  // readOnlyInput={true}
+                  placeholder="Start" // readOnlyInput={true}
                   minDate={moment().toDate()}
                   panelClassName={"calendar-overrideStart"}
                 />
@@ -353,8 +352,7 @@ class UpdateMeeting extends Component {
                   onChange={this.changeHandler}
                   inputClassName="input"
                   className="datePicker"
-                  placeholder="End"
-                  // readOnlyInput={true}
+                  placeholder="End" // readOnlyInput={true}
                   minDate={moment().toDate()}
                   panelClassName={"calendar-overrideEnd"}
                 />
@@ -399,7 +397,9 @@ class UpdateMeeting extends Component {
                 style={{ width: "100%", height: "150px", marginTop: "20px" }}
               >
                 {this.state.invitees.map(invited => {
-                  return <Entry>{invited.displayName}</Entry>;
+                  return (
+                    <Entry key={Math.random()}>{invited.displayName}</Entry>
+                  );
                 })}
               </ScrollPanel>
             </Group>
@@ -423,7 +423,7 @@ class UpdateMeeting extends Component {
                 style={{ width: "100%", height: "75px", marginTop: "20px" }}
               >
                 {this.state.questions.map(question => (
-                  <Entry>{question}</Entry>
+                  <Entry key={Math.random()}>{question}</Entry>
                 ))}
               </ScrollPanel>
             </QGroup>
